Crate nargo

Nargo is the package manager for Noir This name was used because it sounds like cargo and Noir Package Manager abbreviated is npm, which is already taken.

Re-exports

pub use self::errors::NargoError;

pub use self::ops::FuzzExecutionConfig;

pub use self::ops::FuzzFolderConfig;

Modules

constants

errors

foreign_calls

ops

package

workspace

Constants

STACK_SIZE 🔒

Functions

add_unstable_features 🔒

Add any unstable features required by the Package to the Context.

collect_all_files_in_package 🔒

collect_all_files_in_packages_dependencies 🔒

collect_all_files_under_path 🔒

insert_all_files_for_workspace_into_file_manager

insert_all_files_for_workspace_into_file_manager_with_overrides

insert_all_files_into_file_manager 🔒

insert_all_files_under_path

parse_all

prepare_dependencies

prepare_package