nargo
Module foreign_calls

Re-exports

pub use default::DefaultForeignCallBuilder;

pub use default::DefaultForeignCallExecutor;

Modules

default

layers

mocker

print

rpc

transcript

Enums

ForeignCall

This enumeration represents the Brillig foreign calls that are natively supported by nargo. After resolution of a foreign call, nargo will restart execution of the ACVM

ForeignCallError

Traits

ForeignCallExecutor

Interface for executing foreign calls