nargo::foreign_calls::layers
Struct Layer

pub struct Layer<H, I> {
    pub handler: H,
    pub inner: I,
}

Forwards to the inner executor if its own handler doesn’t handle the call.

Fields
handler: Hinner: I
Implementations

impl<H, I> Layer<H, I>

pub fn new(handler: H, inner: I) -> Self

Create a layer from two handlers

impl<H> Layer<H, Empty>

pub fn or_empty(handler: H) -> Self

Create a layer from a handler. If the handler doesn’t handle a call, a default empty response is returned.

impl<H> Layer<H, Unhandled>

pub fn or_unhandled(handler: H) -> Self

Create a layer from a handler. If the handler doesn’t handle a call, NoHandler error is returned.

impl Layer<Unhandled, Unhandled>

pub fn unhandled() -> Self

A base layer that doesn’t handle anything.

impl<H, I> Layer<H, I>

pub fn add_layer<J>(self, handler: J) -> Layer<J, Self>

Add another layer on top of this one.

pub fn handler(&self) -> &H

pub fn inner(&self) -> &I