nargo

Module ops

Modules

check 🔒

compile 🔒

debug

execute 🔒

fuzz 🔒

optimize 🔒

test 🔒

Structs

FuzzConfig

FuzzExecutionConfig

Configuration for fuzzing loop execution

FuzzFolderConfig

Folder configuration for fuzzing

Enums

FuzzingRunStatus

TestStatus

Functions

check_crate_and_report_errors

Run the lexing, parsing, name resolution, and type checking passes and report any warnings and errors found.

check_expected_failure_message

check_program

Run each function through a circuit simulator to check that they are solvable.

collect_errors

Constructs a single CompilationResult for a collection of CompilationResults, merging the set of warnings/errors.

compile_contract

compile_program

compile_program_with_debug_instrumenter

execute_program

Execute a Program, returning the WitnessStack.

execute_program_with_profiling

Execute a Program with profiling turned on, returning the WitnessStack along with the ProfilingSamples.

fuzz_test

Runs the fuzzer on a test function. This assumes the function has arguments.

optimize_contract

optimize_program

report_errors

run_fuzzing_harness

run_or_fuzz_test

Runs a test function. This will either run the test or fuzz it, depending on whether the function has arguments.

run_test

Runs a test function. This assumes the function has no arguments.

test_status_program_compile_fail

Test function failed to compile

test_status_program_compile_pass

The test function compiled successfully.