nargo::foreign_calls
Module default

Structs

DefaultForeignCallBuilder

A builder for DefaultForeignCallLayers where we can enable fields based on feature flags, which is easier than providing different overrides for a new method.

DefaultForeignCallExecutor

Convenience constructor for code that used to create the executor this way.

Type Aliases

DefaultForeignCallLayers

Facilitate static typing of layers on a base layer, so inner layers can be accessed.