nargo::ops
Module test

Structs

FuzzConfig

TestForeignCallExecutor 🔒

A specialized foreign call executor which tracks whether it has encountered any unknown foreign calls

Enums

TestStatus

Functions

check_expected_failure_message

fuzz_test

Runs the fuzzer on a test function. This assumes the function has arguments.

fuzz_test_impl 🔒

run_or_fuzz_test

Runs a test function. This will either run the test or fuzz it, depending on whether the function has arguments.

run_test

Runs a test function. This assumes the function has no arguments.

run_test_impl 🔒

test_status_program_compile_fail

Test function failed to compile

test_status_program_compile_pass

The test function compiled successfully.